Extend HSTS max time and enable preloading
This commit extends the HSTS max time from 30 seconds to 1 year since it has now been proven to work in production. 1 year is the recommended value for this header and is the timespan within which browsers will only attempt to connect using HTTPS. Preloading is also enabled so that anyone visiting the site for the first time will also benefit from this protection.